+++ include/onyx/builtin_shaders.hpp
@@ -37,7 +37,7 @@ in vec3 io_pos;
 
 out vec4 o_color;
 
-uniform vec3 u_camPos;
+uniform vec3 u_cam_pos;
 uniform vec4 u_color;
 
 struct Fog
@@ -55,14 +55,14 @@ void main()
 
 	if (!u_fog.enabled) return;
 
-	float camDist = distance(u_camPos, io_pos);
+	float cam_dist = distance(u_cam_pos, io_pos);
 
-	if (camDist > u_fog.start)
+	if (cam_dist > u_fog.start)
 	{
-		float fogFactor = (camDist - u_fog.start) / (u_fog.end - u_fog.start);
-		fogFactor = clamp(fogFactor, 0.0, 1.0);
+		float fog_factor = (cam_dist - u_fog.start) / (u_fog.end - u_fog.start);
+		fog_factor = clamp(fog_factor, 0.0, 1.0);
 		float a = o_color.a;
-		o_color = mix(o_color, vec4(u_fog.color, 1.0), fogFactor);
+		o_color = mix(o_color, vec4(u_fog.color, 1.0), fog_factor);
 		o_color.a = a;
 	}
 }
@@ -124,7 +124,7 @@ in vec4 io_color;
 
 out vec4 o_color;
 
-uniform vec3 u_camPos;
+uniform vec3 u_cam_pos;
 
 struct Fog
 {
@@ -141,14 +141,14 @@ void main()
 
 	if (!u_fog.enabled) return;
 
-	float camDist = distance(u_camPos, io_pos);
+	float cam_dist = distance(u_cam_pos, io_pos);
 
-	if (camDist > u_fog.start)
+	if (cam_dist > u_fog.start)
 	{
-		float fogFactor = (camDist - u_fog.start) / (u_fog.end - u_fog.start);
-		fogFactor = clamp(fogFactor, 0.0, 1.0);
+		float fog_factor = (cam_dist - u_fog.start) / (u_fog.end - u_fog.start);
+		fog_factor = clamp(fog_factor, 0.0, 1.0);
 		float a = o_color.a;
-		o_color = mix(o_color, vec4(u_fog.color, 1.0), fogFactor);
+		o_color = mix(o_color, vec4(u_fog.color, 1.0), fog_factor);
 		o_color.a = a;
 	}
 })ONYXGLSL";
@@ -182,7 +182,7 @@ in vec4 io_color;
 
 out vec4 o_color;
 
-uniform vec3 u_camPos;
+uniform vec3 u_cam_pos;
 
 struct Fog
 {
@@ -198,14 +198,14 @@ void main() {
 
 	if (!u_fog.enabled) return;
 
-	float camDist = distance(u_camPos, io_pos);
+	float cam_dist = distance(u_cam_pos, io_pos);
 
-	if (camDist > u_fog.start)
+	if (cam_dist > u_fog.start)
 	{
-		float fogFactor = (camDist - u_fog.start) / (u_fog.end - u_fog.start);
-		fogFactor = clamp(fogFactor, 0.0, 1.0);
+		float fog_factor = (cam_dist - u_fog.start) / (u_fog.end - u_fog.start);
+		fog_factor = clamp(fog_factor, 0.0, 1.0);
 		float a = o_color.a;
-		o_color = mix(o_color, vec4(u_fog.color, 1.0), fogFactor);
+		o_color = mix(o_color, vec4(u_fog.color, 1.0), fog_factor);
 		o_color.a = a;
 	}
 }
@@ -217,11 +217,11 @@ void main() {
 
 layout (location = 0) in vec3 i_pos;
 layout (location = 1) in vec4 i_color;
-layout (location = 2) in vec2 i_texCoord;
+layout (location = 2) in vec2 i_tex_coord;
 
 out vec3 io_pos;
 out vec4 io_color;
-out vec2 io_texCoord;
+out vec2 io_tex_coord;
 
 uniform mat4 u_model;
 uniform mat4 u_view;
@@ -232,7 +232,7 @@ void main()
 	gl_Position = u_projection * u_view * u_model * vec4(i_pos, 1.0);
 	io_pos = vec3(u_model * vec4(i_pos, 1.0));
 	io_color = i_color;
-	io_texCoord = i_texCoord;
+	io_tex_coord = i_tex_coord;
 }
 
 // ------------------------------------------------------------------------
@@ -242,11 +242,11 @@ void main()
 
 in vec3 io_pos;
 in vec4 io_color;
-in vec2 io_texCoord;
+in vec2 io_tex_coord;
 
 out vec4 o_color;
 
-uniform vec3 u_camPos;
+uniform vec3 u_cam_pos;
 uniform sampler2D u_tex;
 
 struct Fog
@@ -260,18 +260,18 @@ uniform Fog u_fog;
 
 void main()
 {
-	o_color = io_color * texture(u_tex, io_texCoord);
+	o_color = io_color * texture(u_tex, io_tex_coord);
 
 	if (!u_fog.enabled) return;
 
-	float camDist = distance(u_camPos, io_pos);
+	float cam_dist = distance(u_cam_pos, io_pos);
 
-	if (camDist > u_fog.start)
+	if (cam_dist > u_fog.start)
 	{
-		float fogFactor = (camDist - u_fog.start) / (u_fog.end - u_fog.start);
-		fogFactor = clamp(fogFactor, 0.0, 1.0);
+		float fog_factor = (cam_dist - u_fog.start) / (u_fog.end - u_fog.start);
+		fog_factor = clamp(fog_factor, 0.0, 1.0);
 		float a = o_color.a;
-		o_color = mix(o_color, vec4(u_fog.color, 1.0), fogFactor);
+		o_color = mix(o_color, vec4(u_fog.color, 1.0), fog_factor);
 		o_color.a = a;
 	}
 }
@@ -288,7 +288,7 @@ out vec3 io_pos;
 out vec4 io_color;
 
 uniform mat4 u_model;
-uniform mat4 u_inverseModel;
+uniform mat4 u_inverse_model;
 uniform mat4 u_view;
 uniform mat4 u_projection;
 
@@ -298,7 +298,7 @@ struct Lighting
 {
 	bool enabled;
 	vec3 color;
-	float ambientStrength;
+	float ambient_strength;
 	vec3 direction;
 };
 
@@ -314,13 +314,13 @@ void main()
 		return;
 	}
 
-	vec3 normal = normalize(mat3(transpose(u_inverseModel)) * i_normal);
+	vec3 normal = normalize(mat3(transpose(u_inverse_model)) * i_normal);
 	vec3 color = u_lighting.color * u_color.rgb;
-	vec3 ambient = color * u_lighting.ambientStrength;
+	vec3 ambient = color * u_lighting.ambient_strength;
 	
-	vec3 lightDir = normalize(-u_lighting.direction);
-	float diffuseFactor = max(dot(normal, lightDir), 0.0);
-	vec3 diffuse = color * diffuseFactor;
+	vec3 light_dir = normalize(-u_lighting.direction);
+	float diffuse_factor = max(dot(normal, light_dir), 0.0);
+	vec3 diffuse = color * diffuse_factor;
 
 	io_color = vec4(ambient + diffuse, u_color.a);
 }
@@ -335,7 +335,7 @@ out vec4 o_color;
 in vec3 io_pos;
 in vec4 io_color;
 
-uniform vec3 u_camPos;
+uniform vec3 u_cam_pos;
 
 struct Fog
 {
@@ -352,14 +352,14 @@ void main()
 
 	if (!u_fog.enabled) return;
 
-	float camDist = distance(u_camPos, io_pos);
+	float cam_dist = distance(u_cam_pos, io_pos);
 
-	if (camDist > u_fog.start)
+	if (cam_dist > u_fog.start)
 	{
-		float fogFactor = (camDist - u_fog.start) / (u_fog.end - u_fog.start);
-		fogFactor = clamp(fogFactor, 0.0, 1.0);
+		float fog_factor = (cam_dist - u_fog.start) / (u_fog.end - u_fog.start);
+		fog_factor = clamp(fog_factor, 0.0, 1.0);
 		float a = o_color.a;
-		o_color = mix(o_color, vec4(u_fog.color, 1.0), fogFactor);
+		o_color = mix(o_color, vec4(u_fog.color, 1.0), fog_factor);
 		o_color.a = a;
 	}
 }
@@ -375,7 +375,7 @@ out vec3 io_pos;
 out vec4 io_color;
 
 uniform mat4 u_model;
-uniform mat4 u_inverseModel;
+uniform mat4 u_inverse_model;
 uniform mat4 u_view;
 uniform mat4 u_projection;
 
@@ -383,7 +383,7 @@ struct Lighting
 {
 	bool enabled;
 	vec3 color;
-	float ambientStrength;
+	float ambient_strength;
 	vec3 direction;
 };
 
@@ -396,13 +396,13 @@ void main()
 	io_color = vec4(io_pos, 1.0);
 	if (!u_lighting.enabled) return;
 
-	vec3 normal = normalize(mat3(transpose(u_inverseModel)) * i_normal);
+	vec3 normal = normalize(mat3(transpose(u_inverse_model)) * i_normal);
 	vec3 color = u_lighting.color * io_color.rgb;
-	vec3 ambient = color * u_lighting.ambientStrength;
+	vec3 ambient = color * u_lighting.ambient_strength;
 	
-	vec3 lightDir = normalize(-u_lighting.direction);
-	float diffuseFactor = max(dot(normal, lightDir), 0.0);
-	vec3 diffuse = color * diffuseFactor;
+	vec3 light_dir = normalize(-u_lighting.direction);
+	float diffuse_factor = max(dot(normal, light_dir), 0.0);
+	vec3 diffuse = color * diffuse_factor;
 
 	io_color = vec4(ambient + diffuse, io_color.a);
 }
@@ -417,7 +417,7 @@ out vec4 o_color;
 in vec3 io_pos;
 in vec4 io_color;
 
-uniform vec3 u_camPos;
+uniform vec3 u_cam_pos;
 
 struct Fog
 {
@@ -434,14 +434,14 @@ void main()
 
 	if (!u_fog.enabled) return;
 
-	float camDist = distance(u_camPos, io_pos);
+	float cam_dist = distance(u_cam_pos, io_pos);
 
-	if (camDist > u_fog.start)
+	if (cam_dist > u_fog.start)
 	{
-		float fogFactor = (camDist - u_fog.start) / (u_fog.end - u_fog.start);
-		fogFactor = clamp(fogFactor, 0.0, 1.0);
+		float fog_factor = (cam_dist - u_fog.start) / (u_fog.end - u_fog.start);
+		fog_factor = clamp(fog_factor, 0.0, 1.0);
 		float a = o_color.a;
-		o_color = mix(o_color, vec4(u_fog.color, 1.0), fogFactor);
+		o_color = mix(o_color, vec4(u_fog.color, 1.0), fog_factor);
 		o_color.a = a;
 	}
 }
@@ -458,7 +458,7 @@ out vec3 io_pos;
 out vec4 io_color;
 
 uniform mat4 u_model;
-uniform mat4 u_inverseModel;
+uniform mat4 u_inverse_model;
 uniform mat4 u_view;
 uniform mat4 u_projection;
 
@@ -466,7 +466,7 @@ struct Lighting
 {
 	bool enabled;
 	vec3 color;
-	float ambientStrength;
+	float ambient_strength;
 	vec3 direction;
 };
 
@@ -482,13 +482,13 @@ void main()
 		return;
 	}
 
-	vec3 normal = normalize(mat3(transpose(u_inverseModel)) * i_normal);
+	vec3 normal = normalize(mat3(transpose(u_inverse_model)) * i_normal);
 	vec3 color = u_lighting.color * i_color.rgb;
-	vec3 ambient = color * u_lighting.ambientStrength;
+	vec3 ambient = color * u_lighting.ambient_strength;
 	
-	vec3 lightDir = normalize(-u_lighting.direction);
-	float diffuseFactor = max(dot(normal, lightDir), 0.0);
-	vec3 diffuse = color * diffuseFactor;
+	vec3 light_dir = normalize(-u_lighting.direction);
+	float diffuse_factor = max(dot(normal, light_dir), 0.0);
+	vec3 diffuse = color * diffuse_factor;
 
 	io_color = vec4(ambient + diffuse, i_color.a);
 }
@@ -503,7 +503,7 @@ out vec4 o_color;
 in vec3 io_pos;
 in vec4 io_color;
 
-uniform vec3 u_camPos;
+uniform vec3 u_cam_pos;
 
 struct Fog
 {
@@ -520,14 +520,14 @@ void main()
 
 	if (!u_fog.enabled) return;
 
-	float camDist = distance(u_camPos, io_pos);
+	float cam_dist = distance(u_cam_pos, io_pos);
 
-	if (camDist > u_fog.start)
+	if (cam_dist > u_fog.start)
 	{
-		float fogFactor = (camDist - u_fog.start) / (u_fog.end - u_fog.start);
-		fogFactor = clamp(fogFactor, 0.0, 1.0);
+		float fog_factor = (cam_dist - u_fog.start) / (u_fog.end - u_fog.start);
+		fog_factor = clamp(fog_factor, 0.0, 1.0);
 		float a = o_color.a;
-		o_color = mix(o_color, vec4(u_fog.color, 1.0), fogFactor);
+		o_color = mix(o_color, vec4(u_fog.color, 1.0), fog_factor);
 		o_color.a = a;
 	}
 }
@@ -539,16 +539,16 @@ void main()
 
 layout (location = 0) in vec3 i_pos;
 layout (location = 1) in vec4 i_color;
-layout (location = 2) in vec2 i_texCoord;
+layout (location = 2) in vec2 i_tex_coord;
 layout (location = 3) in vec3 i_normal;
 
 out vec3 io_pos;
-out float io_diffuseFactor;
+out float io_diffuse_factor;
 out vec4 io_color;
-out vec2 io_texCoord;
+out vec2 io_tex_coord;
 
 uniform mat4 u_model;
-uniform mat4 u_inverseModel;
+uniform mat4 u_inverse_model;
 uniform mat4 u_view;
 uniform mat4 u_projection;
 
@@ -556,7 +556,7 @@ struct Lighting
 {
 	bool enabled;
 	vec3 color;
-	float ambientStrength;
+	float ambient_strength;
 	vec3 direction;
 };
 
@@ -565,19 +565,19 @@ uniform Lighting u_lighting;
 void main()
 {
 	gl_Position = u_projection * u_view * u_model * vec4(i_pos, 1.0);
-	io_texCoord = i_texCoord;
+	io_tex_coord = i_tex_coord;
 	io_color = i_color;
 	io_pos = vec3(u_model * vec4(i_pos, 1.0));
 
 	if (!u_lighting.enabled)
 	{
-		io_diffuseFactor = 1.0;
+		io_diffuse_factor = 1.0;
 		return;
 	}
 
-	vec3 normal = normalize(mat3(transpose(u_inverseModel)) * i_normal);
-	vec3 lightDir = normalize(-u_lighting.direction);
-	io_diffuseFactor = max(dot(normal, lightDir), 0.0);
+	vec3 normal = normalize(mat3(transpose(u_inverse_model)) * i_normal);
+	vec3 light_dir = normalize(-u_lighting.direction);
+	io_diffuse_factor = max(dot(normal, light_dir), 0.0);
 }
 
 // ------------------------------------------------------------------------
@@ -588,18 +588,18 @@ void main()
 out vec4 o_color;
 
 in vec3 io_pos;
-in float io_diffuseFactor;
+in float io_diffuse_factor;
 in vec4 io_color;
-in vec2 io_texCoord;
+in vec2 io_tex_coord;
 
-uniform vec3 u_camPos;
+uniform vec3 u_cam_pos;
 uniform sampler2D u_tex;
 
 struct Lighting
 {
 	bool enabled;
 	vec3 color;
-	float ambientStrength;
+	float ambient_strength;
 	vec3 direction;
 };
 
@@ -615,42 +615,42 @@ uniform Fog u_fog;
 
 void main()
 {
-	vec4 texColor = texture(u_tex, io_texCoord);
+	vec4 texColor = texture(u_tex, io_tex_coord);
 
 	if (!u_lighting.enabled)
 	{
 		o_color = io_color * texColor;
 		if (!u_fog.enabled) return;
 
-		float camDist = distance(u_camPos, io_pos);
+		float cam_dist = distance(u_cam_pos, io_pos);
 
-		if (camDist > u_fog.start)
+		if (cam_dist > u_fog.start)
 		{
-			float fogFactor = (camDist - u_fog.start) / (u_fog.end - u_fog.start);
-			fogFactor = clamp(fogFactor, 0.0, 1.0);
+			float fog_factor = (cam_dist - u_fog.start) / (u_fog.end - u_fog.start);
+			fog_factor = clamp(fog_factor, 0.0, 1.0);
 			float a = o_color.a;
-			o_color = mix(o_color, vec4(u_fog.color, 1.0), fogFactor);
+			o_color = mix(o_color, vec4(u_fog.color, 1.0), fog_factor);
 			o_color.a = a;
 		}
 	}
 
 	vec3 color = u_lighting.color * io_color.rgb * texColor.rgb;
 
-	vec3 ambient = color * u_lighting.ambientStrength;
-	vec3 diffuse = color * io_diffuseFactor;
+	vec3 ambient = color * u_lighting.ambient_strength;
+	vec3 diffuse = color * io_diffuse_factor;
 	
 	o_color = vec4(diffuse + ambient, texColor.a * io_color.a);
 
 	if (!u_fog.enabled) return;
 
-	float camDist = distance(u_camPos, io_pos);
+	float cam_dist = distance(u_cam_pos, io_pos);
 
-	if (camDist > u_fog.start)
+	if (cam_dist > u_fog.start)
 	{
-		float fogFactor = (camDist - u_fog.start) / (u_fog.end - u_fog.start);
-		fogFactor = clamp(fogFactor, 0.0, 1.0);
+		float fog_factor = (cam_dist - u_fog.start) / (u_fog.end - u_fog.start);
+		fog_factor = clamp(fog_factor, 0.0, 1.0);
 		float a = o_color.a;
-		o_color = mix(o_color, vec4(u_fog.color, 1.0), fogFactor);
+		o_color = mix(o_color, vec4(u_fog.color, 1.0), fog_factor);
 		o_color.a = a;
 	}
 }
@@ -660,15 +660,15 @@ void main()
 #version 410 core
 
 layout (location = 0) in vec3 i_pos;
-layout (location = 2) in vec2 i_texCoord;
+layout (location = 2) in vec2 i_tex_coord;
 layout (location = 3) in vec3 i_normal;
 
 out vec3 io_pos;
-out float io_diffuseFactor;
-out vec2 io_texCoord;
+out float io_diffuse_factor;
+out vec2 io_tex_coord;
 
 uniform mat4 u_model;
-uniform mat4 u_inverseModel;
+uniform mat4 u_inverse_model;
 uniform mat4 u_view;
 uniform mat4 u_projection;
 
@@ -676,7 +676,7 @@ struct Lighting
 {
 	bool enabled;
 	vec3 color;
-	float ambientStrength;
+	float ambient_strength;
 	vec3 direction;
 };
 
@@ -686,17 +686,17 @@ void main()
 {
 	gl_Position = u_projection * u_view * u_model * vec4(i_pos, 1.0);
 	io_pos = vec3(u_model * vec4(i_pos, 1.0));
-	io_texCoord = i_texCoord;
+	io_tex_coord = i_tex_coord;
 
 	if (!u_lighting.enabled)
 	{
-		io_diffuseFactor = 1.0;
+		io_diffuse_factor = 1.0;
 		return;
 	}
 
-	vec3 normal = normalize(mat3(transpose(u_inverseModel)) * i_normal);
-	vec3 lightDir = normalize(-u_lighting.direction);
-	io_diffuseFactor = max(dot(normal, lightDir), 0.0);
+	vec3 normal = normalize(mat3(transpose(u_inverse_model)) * i_normal);
+	vec3 light_dir = normalize(-u_lighting.direction);
+	io_diffuse_factor = max(dot(normal, light_dir), 0.0);
 }
 
 // ------------------------------------------------------------------------
@@ -707,17 +707,17 @@ void main()
 out vec4 o_color;
 
 in vec3 io_pos;
-in float io_diffuseFactor;
-in vec2 io_texCoord;
+in float io_diffuse_factor;
+in vec2 io_tex_coord;
 
-uniform vec3 u_camPos;
+uniform vec3 u_cam_pos;
 uniform sampler2D u_tex;
 
 struct Lighting
 {
 	bool enabled;
 	vec3 color;
-	float ambientStrength;
+	float ambient_strength;
 	vec3 direction;
 };
 
@@ -733,38 +733,38 @@ uniform Fog u_fog;
 
 void main()
 {
-	vec4 texColor = texture(u_tex, io_texCoord);
+	vec4 texColor = texture(u_tex, io_tex_coord);
 
 	if (!u_lighting.enabled)
 	{
 		o_color = texColor;
 		if (!u_fog.enabled) return;
 
-		float camDist = distance(u_camPos, io_pos);
+		float cam_dist = distance(u_cam_pos, io_pos);
 
-		if (camDist > u_fog.start)
+		if (cam_dist > u_fog.start)
 		{
-			float fogFactor = (camDist - u_fog.start) / (u_fog.end - u_fog.start);
-			fogFactor = clamp(fogFactor, 0.0, 1.0);
-			o_color = mix(o_color, vec4(u_fog.color, 1.0), fogFactor);
+			float fog_factor = (cam_dist - u_fog.start) / (u_fog.end - u_fog.start);
+			fog_factor = clamp(fog_factor, 0.0, 1.0);
+			o_color = mix(o_color, vec4(u_fog.color, 1.0), fog_factor);
 		}
 	}
 
 	vec3 color = u_lighting.color * texColor.rgb;
-	vec3 ambient = color * u_lighting.ambientStrength;
-	vec3 diffuse = color * io_diffuseFactor;
+	vec3 ambient = color * u_lighting.ambient_strength;
+	vec3 diffuse = color * io_diffuse_factor;
 	o_color = vec4(diffuse + ambient, texColor.a);
 	
 	if (!u_fog.enabled) return;
 
-	float camDist = distance(u_camPos, io_pos);
+	float cam_dist = distance(u_cam_pos, io_pos);
 
-	if (camDist > u_fog.start)
+	if (cam_dist > u_fog.start)
 	{
-		float fogFactor = (camDist - u_fog.start) / (u_fog.end - u_fog.start);
-		fogFactor = clamp(fogFactor, 0.0, 1.0);
+		float fog_factor = (cam_dist - u_fog.start) / (u_fog.end - u_fog.start);
+		fog_factor = clamp(fog_factor, 0.0, 1.0);
 		float a = o_color.a;
-		o_color = mix(o_color, vec4(u_fog.color, 1.0), fogFactor);
+		o_color = mix(o_color, vec4(u_fog.color, 1.0), fog_factor);
 		o_color.a = a;
 	}
 }
@@ -774,10 +774,10 @@ void main()
 #version 410 core
 
 layout (location = 0) in vec3 i_pos;
-layout (location = 2) in vec2 i_texCoord;
+layout (location = 2) in vec2 i_tex_coord;
 
 out vec3 io_pos;
-out vec2 io_texCoord;
+out vec2 io_tex_coord;
 
 uniform mat4 u_model;
 uniform mat4 u_view;
@@ -786,7 +786,7 @@ uniform mat4 u_projection;
 void main()
 {
 	gl_Position = u_projection * u_view * u_model * vec4(i_pos, 1.0);
-	io_texCoord = i_texCoord;
+	io_tex_coord = i_tex_coord;
 	io_pos = vec3(u_model * vec4(i_pos, 1.0));
 }
 
@@ -796,11 +796,11 @@ void main()
 #version 410 core
 
 in vec3 io_pos;
-in vec2 io_texCoord;
+in vec2 io_tex_coord;
 
 out vec4 o_color;
 
-uniform vec3 u_camPos;
+uniform vec3 u_cam_pos;
 uniform sampler2D u_tex;
 
 struct Fog
@@ -814,18 +814,18 @@ uniform Fog u_fog;
 
 void main()
 {
-	o_color = texture(u_tex, io_texCoord);
+	o_color = texture(u_tex, io_tex_coord);
 	
 	if (!u_fog.enabled) return;
 
-	float camDist = distance(u_camPos, io_pos);
+	float cam_dist = distance(u_cam_pos, io_pos);
 
-	if (camDist > u_fog.start)
+	if (cam_dist > u_fog.start)
 	{
-		float fogFactor = (camDist - u_fog.start) / (u_fog.end - u_fog.start);
-		fogFactor = clamp(fogFactor, 0.0, 1.0);
+		float fog_factor = (cam_dist - u_fog.start) / (u_fog.end - u_fog.start);
+		fog_factor = clamp(fog_factor, 0.0, 1.0);
 		float a = o_color.a;
-		o_color = mix(o_color, vec4(u_fog.color, 1.0), fogFactor);
+		o_color = mix(o_color, vec4(u_fog.color, 1.0), fog_factor);
 		o_color.a = a;
 	}
 }
@@ -836,9 +836,9 @@ void main()
 #version 410 core
 
 layout (location = 0) in vec3 i_pos;
-layout (location = 2) in vec2 i_texCoord;
+layout (location = 2) in vec2 i_tex_coord;
 
-out vec2 io_texCoord;
+out vec2 io_tex_coord;
 
 uniform mat4 u_model;
 uniform mat4 u_projection;
@@ -846,7 +846,7 @@ uniform mat4 u_projection;
 void main()
 {
 	gl_Position = u_projection * u_model * vec4(i_pos, 1.0f);
-	io_texCoord = i_texCoord;
+	io_tex_coord = i_tex_coord;
 }
 
 // ------------------------------------------------------------------------
@@ -854,7 +854,7 @@ void main()
 
 #version 410 core
 
-in vec2 io_texCoord;
+in vec2 io_tex_coord;
 
 out vec4 o_color;
 
@@ -862,7 +862,7 @@ uniform sampler2D u_tex;
 
 void main()
 {
-	o_color = texture(u_tex, io_texCoord);
+	o_color = texture(u_tex, io_tex_coord);
 }
 
 )ONYXGLSL";
@@ -872,7 +872,7 @@ void main()
 
 layout (location = 0) in vec4 i_vertex;
 
-out vec2 io_texCoord;
+out vec2 io_tex_coord;
 out vec3 io_pos;
 
 uniform mat4 u_model;
@@ -882,7 +882,7 @@ uniform mat4 u_projection;
 void main()
 {
 	gl_Position = u_projection * u_view * u_model * vec4(i_vertex.xy, 0.0, 1.0);
-	io_texCoord = i_vertex.zw;
+	io_tex_coord = i_vertex.zw;
 	io_pos = vec3(u_model * vec4(i_vertex.xy, 0.0, 1.0));
 }
 
@@ -891,14 +891,14 @@ void main()
 
 #version 410 core
 
-in vec2 io_texCoord;
+in vec2 io_tex_coord;
 in vec3 io_pos;
 
 out vec4 o_color;
 
 uniform sampler2D u_tex;
 uniform vec4 u_color;
-uniform vec3 u_camPos;
+uniform vec3 u_cam_pos;
 
 struct Fog
 {
@@ -911,17 +911,17 @@ uniform Fog u_fog;
 
 void main()
 {
-	o_color = u_color * vec4(1.0, 1.0, 1.0, texture(u_tex, io_texCoord).r);
+	o_color = u_color * vec4(1.0, 1.0, 1.0, texture(u_tex, io_tex_coord).r);
 	if (!u_fog.enabled) return;
 
-	float camDist = distance(u_camPos, io_pos);
+	float cam_dist = distance(u_cam_pos, io_pos);
 
-	if (camDist > u_fog.start)
+	if (cam_dist > u_fog.start)
 	{
-		float fogFactor = (camDist - u_fog.start) / (u_fog.end - u_fog.start);
-		fogFactor = clamp(fogFactor, 0.0, 1.0);
+		float fog_factor = (cam_dist - u_fog.start) / (u_fog.end - u_fog.start);
+		fog_factor = clamp(fog_factor, 0.0, 1.0);
 		float a = o_color.a;
-		o_color = mix(o_color, vec4(u_fog.color, 1.0), fogFactor);
+		o_color = mix(o_color, vec4(u_fog.color, 1.0), fog_factor);
 		o_color.a = a;
 	}
 }
@@ -932,7 +932,7 @@ void main()
 
 layout (location = 0) in vec4 i_vertex;
 
-out vec2 io_texCoord;
+out vec2 io_tex_coord;
 
 uniform mat4 u_model;
 uniform mat4 u_projection;
@@ -940,7 +940,7 @@ uniform mat4 u_projection;
 void main()
 {
 	gl_Position = u_projection * u_model * vec4(i_vertex.xy, 0.0, 1.0);
-	io_texCoord = i_vertex.zw;
+	io_tex_coord = i_vertex.zw;
 }
 
 // ------------------------------------------------------------------------
@@ -948,7 +948,7 @@ void main()
 
 #version 410 core
 
-in vec2 io_texCoord;
+in vec2 io_tex_coord;
 
 out vec4 o_color;
 
@@ -957,7 +957,7 @@ uniform vec4 u_color;
 
 void main()
 {
-	o_color = u_color * vec4(1.0, 1.0, 1.0, texture(u_tex, io_texCoord).r);
+	o_color = u_color * vec4(1.0, 1.0, 1.0, texture(u_tex, io_tex_coord).r);
 }
 
 )ONYXGLSL";