+++ include/onyx/prelude.hpp
@@ -1,0 +1,11 @@
+#pragma once
+
+#include <onyx/core.hpp>
+#include <onyx/window.hpp>
+#include <onyx/input_handler.hpp>
+#include <onyx/camera.hpp>
+#include <onyx/renderable.hpp>
+#include <onyx/math_wrappers.hpp>
+#include <onyx/renderable.hpp>
+
+using onyx::math::Vec3, onyx::math::Vec4;