+++ src/text_renderable_3d.cpp
@@ -231,7 +231,7 @@ void onyx::TextRenderable3D::render(const Mat4& view, const Mat4& proj, const Ve
 	this->shader.set_mat4("u_model", this->model);
 	this->shader.set_mat4("u_view", view);
 	this->shader.set_mat4("u_projection", proj);
-	this->shader.set_vec3("u_camPos", cam_pos);
+	this->shader.set_vec3("u_cam_pos", cam_pos);
 	for (CharRenderable& c : this->chars) c.render();
 }